requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-dotenv==1.0.1
PyYAML==6.0.2
pandas==2.2.3
orjson==3.10.12
python-dateutil==2.9.0.post0
pytest==8.3.4
agno==2.5.10
